core: build the library and the suite without warnings
Four warnings at the start of this work, none of them acted on, and one
of them was a real defect that had been shipping -- the elision marker
emitting a broken bar, fixed in 0a3dd85 on the strength of exactly the
warning nobody had read. That is the argument for the other three.

Two are the same mismatch. CellImage::key is quint64 and
QPixmap::cacheKey() returns qint64, so the aggregate initialiser
narrowed and the test's comparison mixed signs. The field is unsigned on
purpose: the value is an opaque handle, compared and used as a map key,
never ordered and never arithmetic. So the conversion is written out at
both call sites rather than the field being changed, and the header says
why -- the intent is now in the source instead of in a warning the
reader has to reconstruct.

The third is dead code: cubeLevel() mapped an RGB component to a cube
level for the nearest-match quantiser, and matching in CIELAB compares
whole colours against the candidate table instead of choosing a level
per channel. Nothing was left to invert.

A clean build of the library and of the suite is now silent, which is
worth more than the three fixes: a build with four warnings in it is one
where the fifth arrives unnoticed.

+++ src/render/cell_paint.cpp
@@ -97,7 +97,7 @@ void CellPaintEngine::drawPixmap(const QRectF &r, const QPixmap &pm, const QRect
 	QRect c = toCells(r);
 	if (!c.isValid()) return;
 	if (c.width() >= 2 && c.height() >= 2)      // section 5.7: real image -> placement
-		dev_->placements.append({pm.cacheKey(), c, pm});
+		dev_->placements.append({quint64(pm.cacheKey()), c, pm});
 	else                                        // tiny icon -> glyph substitution (section 8.6)
 		dev_->buffer().text(c.left(), c.top(), QStringLiteral("▒"));
 }
